Update Macalania Woods to do RNG2 Tracking
Updates logic to account for RNG2

Will wait in front of butterfly to see if the sphermiroph weakness is something that we can handle. If it isn't, then we wait for rng2 to cycle to something favorable.

As such, we only will attempt to charge rikku either by accident or on something that we can steal from, and we only go for a steal if we can charge. This should speed up mwoods significantly.