Spherimorph Grenade Skip
We can use RNG26 to predict Rikku's mix damage on Spherimorph. This means, that if we know that the mix will do at least 12k damage, this combined with the RNG2 manipulation of his weakness means that as long as Yuna and Kimarhi have gotten their turn we can skip the grenade entirely.