Distance Logging and Filesave Fix
Adds in watcher for distance logging, as well as initial results file.

The distance log is calculated with the "optimal" movements, meaning no additional farming required for anything and swimmer NEA.

Bugfix to saves so that things get got with the new set of saves and non-nemesis.

Bugfix to reset that could happen on boat, but shouldn't ever happen.